refactor: split billing/usage.ts into tier, metrics, and quota modules
Phase 3 of Break Up God Modules PRD. Split usage.ts (552 lines) into
focused single-responsibility files with a barrel re-export preserving
all existing consumer imports.

Changes:
- Extract billing/tier.ts (32 lines): resolveUserTier, getUserTier
- Extract billing/usage-metrics.ts (323 lines): monthly CRUD, increment/decrement, getUsageSummary
- Extract billing/quota.ts (208 lines): atomicCheckAndIncrementUsage, atomicCheckAndIncrementPdfUsage
- Convert billing/usage.ts to 23-line barrel re-export
- Update prds/break-up-god-modules/todos.md with Phase 3 results

Closes #263, #266, #269
